Move config related test out of bootstrap_test.py
